Update README.md
My 5 websites that I like, along with a few things I don't about them. Also Raydium comes up as a website that is not "safe" however this is a finance crypto defi company that is backed by the foundation who created a product with which they use as a commodity.